requirements.txt
openai>=1.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
argparse>=1.4.0
pathlib>=1.0.1
requests>=2.25.0 